pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gravedigger"
version = "1.0.0"
description = "Resource unpacker/repacker for Dangerous Dave in the Haunted Mansion"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
]

[project.scripts]
gravedigger = "gravedigger.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "mypy>=1.14",
    "nuitka>=2.5",
    "zstandard>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
cache_dir = ".cache/pytest"

[tool.coverage.run]
data_file = ".cache/.coverage"

[tool.ruff]
target-version = "py313"
line-length = 99
cache-dir = ".cache/ruff"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "C4", "PTH", "RET", "PIE", "RUF", "PERF", "FLY"]

[tool.mypy]
strict = true
python_version = "3.13"
cache_dir = ".cache/mypy"
exclude = ['\.cache/', 'reference/']
